Extracted startup script to file

+++ start_backend.sh
@@ -11,8 +11,10 @@ else
   echo "Process was already stopped, no need to stop it again"
 fi
 
-cd app
-chmod 700 backend-1.0.jar.new
-mv --force backend-1.0.jar backend-1.0.jar.bak
-mv --force backend-1.0.jar.new backend-1.0.jar
+if [ -f backend-1.0.jar.new ]; then
+  chmod 700 backend-1.0.jar.new
+  mv --force backend-1.0.jar backend-1.0.jar.bak
+  mv --force backend-1.0.jar.new backend-1.0.jar
+fi
+
 nohup java -jar backend-1.0.jar --spring.profiles.active=aws >> application.log 2>> application.log &
